refactor and credits
