+++ tools/emulator/requirements-simulator.txt
